fix: Reload window if token is invalid

+++ web/src/stores/auth.ts
@@ -34,8 +34,13 @@ export const useAuthStore = defineStore("auth", () => {
         if (user.response.status != 200) {
             console.warn("[auth] Could not login!");
             console.error(user);
-            localStorage.clear();
 
+            if (localStorage.getItem("token")) {
+                localStorage.clear();
+                location.reload();
+            }
+
+            localStorage.clear();
             return false;
         }
